KeyBinding.KeyのテストをKeyTestsに分離
