+++ macSKKTests/CurrentInputTests.swift
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: 2024 mtgto <hogerappa@gmail.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-
 import XCTest
 
 @testable import macSKK
